Refactor component export normalization and import transformation utilities
- Removed transform utilities from `use-vibes` package and moved them to `@vibes.diy/prompts`.
- Updated import paths in various components to use the new `@vibes.diy/prompts` package.
- Deleted obsolete `normalizeComponentExports` utility and integrated its functionality into the new structure.
- Enhanced the `transformImports` function to handle import transformations for core libraries.
- Updated tests to reflect changes in import paths and utility functions.
