+++ prompts/pkg/component-export-transforms.ts
