--- use-vibes/base/hooks/vibes-gen/transformImports.ts
@@ -1,20 +1,0 @@
-export const coreImportMap = [
-  'react',
-  'react-dom',
-  'react-dom/client',
-  'use-fireproof',
-  'call-ai',
-  'use-vibes',
-];
-
-export function transformImports(code: string): string {
-  return code.replace(
-    /import\s+(?:(?:\{[^}]*\}|\*\s+as\s+\w+|\w+)\s+from\s+)?['"]([^/][^'"]*)['"];?/g,
-    (match, importPath) => {
-      if (coreImportMap.includes(importPath)) {
-        return match;
-      }
-      return match.replace(`"${importPath}"`, `"https://esm.sh/${importPath}"`);
-    }
-  );
-}
