+++ use-vibes/pkg/index.ts
@@ -20,10 +20,6 @@ export {
   type UseVibesResult,
   type VibeDocument,
 
-  // Transform utilities
-  transformImports,
-  coreImportMap,
-
   // Type namespaces
   type Fireproof,
   type CallAI,
