+++ vibes.diy/pkg/app/utils/exportHtml.ts
@@ -1,7 +1,9 @@
 import { VibesDiyEnv } from "../config/env.js";
 import iframeTemplateRaw from "../components/ResultPreview/templates/iframe-template.html?raw";
-import { normalizeComponentExports } from "./normalizeComponentExports.js";
-import { transformImports } from "use-vibes";
+import {
+  normalizeComponentExports,
+  transformImports,
+} from "@vibes.diy/prompts";
 
 export function generateStandaloneHtml(params: { code: string }): string {
   const normalized = normalizeComponentExports(params.code);
